логика работы с формой в отдельном компоненте
